+++ frontend/techpick/src/features/nodeManagement/DnD/ui/Folder.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Image from 'next/image';
-import { folderWrapper } from '@/features/DnD/folder.css';
+import { folderWrapper } from '@/features/nodeManagement/DnD/ui/folder.css';
 import { NodeApi } from 'react-arborist';
-import { useDragHook } from '@/features/DnD/hooks/useDragHook';
+import { useDragHook } from '@/features/nodeManagement/DnD/hooks/useDragHook';
 import { useTreeStore } from '@/shared/stores/treeStore';
 
 export function Folder({ node }: { node: NodeApi }) {
