+++ frontend/techpick/src/features/nodeManagement/utils/createNode.ts
