+++ frontend/techpick/src/widgets/LinkEditorSection/LinkEditor.tsx
@@ -1,9 +1,9 @@
 import React, { useCallback, useRef } from 'react';
 
-import { Folder } from '@/features/DnD/Folder';
-import { Pick } from '@/features/DnD/Pick';
+import { Folder } from '@/features/nodeManagement/DnD/ui/Folder';
+import { Pick } from '@/features/nodeManagement/DnD/ui/Pick';
 import { useTreeStore } from '@/shared/stores/treeStore';
-import { useDropHook } from '@/features/DnD/hooks/useDropHook';
+import { useDropHook } from '@/features/nodeManagement/DnD/hooks/useDropHook';
 import {
   folderViewSection,
   linkViewSection,
